refactor: refactor current wallet functionalities
