+++ src/main.rs
@@ -1,13 +1,13 @@
+mod create;
+mod list;
+mod utils;
+
+use crate::create::{create_wallet, DEFAULT_WALLETS_VAULT_PATH};
+use crate::list::{get_next_wallet_index, print_wallet_list};
 use clap::{ArgEnum, Parser, Subcommand};
 use fuels::prelude::*;
 use fuels::signers::wallet::Wallet;
-use std::collections::HashMap;
-use std::io::stdout;
-use std::io::Write;
 use std::path::PathBuf;
-use termion::screen::AlternateScreen;
-
-const DEFAULT_WALLETS_VAULT_PATH: &str = ".fuel/wallets/";
 
 #[derive(Debug, Parser)]
 #[clap(
@@ -35,8 +35,8 @@ enum Command {
         /// The Bip39 phrase to import the wallet from
         phrase: String,
     },
-    /// Lists all wallets stored in ~/.fuel/wallets/.
-    List,
+    /// Lists all wallets stored in `path`, or in `~/.fuel/wallets/`.
+    List { path: Option<String> },
 }
 
 #[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, ArgEnum)]
@@ -47,152 +47,22 @@ enum OutputFormat {
 }
 
 #[tokio::main]
-async fn main() -> anyhow::Result<()> {
+async fn main() -> Result<(), Error> {
     let app = App::parse();
 
-    match app.command {
+    let _ = match app.command {
         Command::New { path } => create_wallet(path).await,
         Command::Import { phrase } => import_wallet(phrase).await,
-        Command::List => list_wallets(),
+        Command::List { path } => {
+            print_wallet_list(path.unwrap_or_else(|| DEFAULT_WALLETS_VAULT_PATH.to_string()))
+        }
     };
-
     Ok(())
 }
 
-async fn create_wallet(path: Option<PathBuf>) {
-    // Generate wallet from mnenomic phrase.
-    let mnemonic = Wallet::generate_mnemonic_phrase(&mut rand::thread_rng(), 12).unwrap();
-    let wallet = Wallet::new_from_mnemonic_phrase(&mnemonic, None).unwrap();
-
-    let password =
-        rpassword::prompt_password("Please enter a password to encrypt this private key: ")
-            .unwrap();
-
-    let confirmation = rpassword::prompt_password("Please confirm your password: ").unwrap();
-
-    if password != confirmation {
-        println!("Passwords do not match -- try again!");
-        std::process::exit(1);
-    }
-
-    // Wallets are created in ~/.fuel/wallets/ following the format below:
-    // <index>_<public_address>/<uuid>.
-    // The index is the wallet's index in the list of wallets.
-    let path = path.unwrap_or_else(|| {
-        let mut path = home::home_dir().unwrap();
-        path.push(DEFAULT_WALLETS_VAULT_PATH);
-        path.push(format!("{}_{}", get_next_wallet_index(), wallet.address()));
-
-        // create directory if it doesn't exist.
-        if !path.exists() {
-            std::fs::create_dir_all(&path).unwrap();
-        }
-
-        path
-    });
-
-    // Encrypt the wallet and store it in the vault.
-    let uuid = wallet.encrypt(path, password).unwrap();
-
-    // Prints to an alternate screen.
-    // This prevents the mnemonic phrase from being printed to the terminal.
-    let mut screen = AlternateScreen::from(stdout());
-    writeln!(screen, "JSON Wallet uuid: {}\n", uuid).unwrap();
-    writeln!(screen, "Wallet public address: {}\n", wallet.address()).unwrap();
-    writeln!(screen, "Wallet mnemonic phrase: {}\n", mnemonic).unwrap();
-    screen.flush().unwrap();
-
-    let mut input = String::new();
-    println!("### Do not share or lose this mnemonic phrase! Press any key to complete. ###");
-    std::io::stdin().read_line(&mut input).unwrap();
-}
-
-async fn import_wallet(phrase: String) {
+async fn import_wallet(phrase: String) -> Result<(), Error> {
     let wallet = Wallet::new_from_mnemonic_phrase(&phrase, None).unwrap();
 
     println!("Wallet imported: {}", wallet.address());
-}
-
-// Walks through ~/.fuel/wallets/ and returns the next index based
-// on the number of wallets in the vault.
-fn get_next_wallet_index() -> usize {
-    let mut path = home::home_dir().unwrap();
-    path.push(DEFAULT_WALLETS_VAULT_PATH);
-
-    let dirs = match std::fs::read_dir(path) {
-        Ok(dirs) => dirs,
-        Err(_) => {
-            return 0;
-        }
-    };
-
-    let mut highest_index = 0;
-    for dir in dirs {
-        let dir = dir.unwrap();
-        let path = dir.path();
-
-        if path.is_dir() {
-            let last_component = path.components().last().unwrap();
-            let split: Vec<&str> = last_component
-                .as_os_str()
-                .to_str()
-                .unwrap()
-                .split('_')
-                .collect();
-            let index = split.first().unwrap().parse::<usize>().unwrap();
-            if index > highest_index {
-                highest_index = index;
-            }
-        }
-    }
-
-    highest_index + 1
-}
-
-fn list_wallets() {
-    let mut path = home::home_dir().unwrap();
-    path.push(DEFAULT_WALLETS_VAULT_PATH);
-
-    // list directories in the path
-    let dirs = match std::fs::read_dir(path) {
-        Ok(dirs) => dirs,
-        Err(_) => {
-            println!("No wallets found.");
-            return;
-        }
-    };
-
-    let mut wallets = HashMap::new();
-
-    println!("Num.         Address.");
-    for dir in dirs {
-        let dir = dir.unwrap();
-        let path = dir.path();
-
-        if path.is_dir() {
-            // get the last component of the path
-            let last_component = path.components().last().unwrap();
-            let split: Vec<String> = last_component
-                .as_os_str()
-                .to_str()
-                .unwrap()
-                .split('_')
-                .map(|s| s.to_string())
-                .collect();
-
-            let index = split.first().unwrap().clone().parse::<usize>().unwrap();
-            let address = split.last().unwrap().clone();
-            let s = format!("[{}].         0x{}", index, address);
-            wallets.insert(index, s);
-        }
-    }
-
-    // Sort the wallets by index.
-    let mut sorted_wallets = wallets.into_iter().collect::<Vec<_>>();
-    sorted_wallets.sort_by(|a, b| a.0.cmp(&b.0));
-
-    // Print the wallets.
-    for wallet in sorted_wallets {
-        println!("{}", wallet.1);
-    }
+    Ok(())
 }
